pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kaniko_wrapper"]

[project]
name = "kaniko-wrapper"
version = "2.0.0.0"
authors = [
    {name = "Marry", email = "okomarova@saber.games"},
    {name = "EpicMorg", email = "developer@epicm.org"}
]

maintainers = [
    {name = "Marry", email = "okomarova@saber.games"},
    {name = "EpicMorg", email = "developer@epicm.org"}
]
dependencies = [
  "PyYAML",
  "python-dotenv",
  "click",
  "pytest",
  "termcolor",
  "pydantic_settings",
  "colorlog"
]
description = "EpicMorg: Kaniko-Compose Wrapper - CLI wrapper for Kaniko build system"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
keywords = ["kaniko", "docker", "build"]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix"
]

[project.scripts]
kaniko-wrapper = "kaniko_wrapper.kaniko_wrapper:main"

[project.urls]
Homepage = "https://github.com/EpicMorg/kaniko-wrapper"
Documentation = "https://github.com/EpicMorg/kaniko-wrapper/blob/master/README.md"
Repository = "https://github.com/EpicMorg/kaniko-wrapper.git"
"Bug Tracker" = "https://github.com/EpicMorg/kaniko-wrapper/issues"
Changelog = "https://github.com/EpicMorg/kaniko-wrapper/blob/master/CHANGELOG.md"